+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html),
 with the exception that 0.x versions can break between minor versions.
 
+## Unreleased
+### Added
+- Include string snippet in errors for unknown group and invalid escape
+  to make it easier to identify the problem.
+
 ## [0.3.4] - 2020-04-28
 ### Added
 - Support comments using `(?# comment)` syntax